refactor(test): better arrangement in tests in central folder
